# apps/studio/data/vela/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/vela"
    }
  ],
  "paths": {
...
    "/organizations/": {
      "get": {
        "summary": "Organizations:List",
        "operationId": "organizations:list",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Organization"
                  },
                  "title": "Response Organizations List Organizations  Get"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Organizations:Create",
        "operationId": "organizations:create",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "response",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "empty",
                "full"
              ],
              "type": "string",
              "default": "empty",
              "title": "Response"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Organization"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Organizations Create Organizations  Post"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "URL of the created item",
                "schema": {
                  "type": "string"
                }
              }
            },
            "links": {
              "detail": {
                "operationId": "organizations:detail",
                "parameters": {
                  "organization_id": "$response.header.Location#regex:/organizations/(.+)/"
                }
              },
              "update": {
                "operationId": "organizations:update",
                "parameters": {
                  "organization_id": "$response.header.Location#regex:/organizations/(.+)/"
                }
              },
              "delete": {
                "operationId": "organizations:delete",
                "parameters": {
                  "organization_id": "$response.header.Location#regex:/organizations/(.+)/"
                }
              },
              "create_project": {
                "operationId": "organizations:projects:create",
                "parameters": {
                  "organization_id": "$response.header.Location#regex:/organizations/(.+)/"
                }
              },
              "list_projects": {
                "operationId": "organizations:projects:list",
                "parameters": {
                  "organization_id": "$response.header.Location#regex:/organizations/(.+)/"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPError": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "HTTPError"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Organization": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 26,
            "minLength": 26,
            "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
            "format": "ulid",
            "title": "ULID",
            "description": "A ULID (Universally Unique Lexicographically Sortable Identifier)",
            "examples": [
              "01ARZ3NDEKTSV4RRFFQ69G5FAV",
              "01H945P9C3K2QJ8F7N6M4R2E8V"
            ]
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "pattern": "^[^\\x00]*$",
            "title": "Name"
          },
          "locked": {
            "type": "boolean",
            "title": "Locked",
            "default": false
          },
          "require_mfa": {
            "type": "boolean",
            "title": "Require Mfa",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Organization"
      },
...
      "OrganizationCreate": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "pattern": "^[^\\x00]*$",
            "title": "Name"
          },
          "locked": {
            "type": "boolean",
            "title": "Locked",
            "default": false
          },
          "require_mfa": {
            "type": "boolean",
            "title": "Require Mfa",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "OrganizationCreate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}